feat: encode parity with a SIMD codec, named per slice
The old codec cost more to encode parity than the sender spent handing a
shard to the kernel: 5220ns a block against 4620ns. Through the window it
now costs 610ns a block, of which 546ns is the parity itself.

The new codec builds every recovery shard from the number of them a slice
was asked for, where the old one built each independently. So a slice
sealed while the transfer carried four shards cannot be reconstructed
against a decoder expecting eight: it returns the wrong bytes and reports
no error. The suite already pinned the property the old codec had, which
is what caught this.

A shard therefore names the parity its own slice carries, in the header
byte held back for exactly this, and the receiver reconstructs against
that width. A slice keeps the width it was created with, so that every
shard of it can name the same one, which means covering more or less
takes effect on the slice after the one being filled. Nothing waits on
it: the rate controller reacts per tick regardless.

That width arrives from the wire, so it is checked before it reaches the
codec. A width the transfer never announced, or a parity slot lying
outside the width its own shard names, would otherwise panic the
receiver, and a width that merely differed from the true one would
reconstruct the wrong bytes silently. A slice also keeps the first width
it was told rather than the last, so nothing already stored under one
width is reconstructed against another, and the widest parity a transfer
allows is built when the transfer is joined, so an announcement the
codec cannot serve is refused there.

Widths are capped at 255 by the field, which costs nothing: the old codec
could not hold more than 255 shards of any kind. Parity counts narrow to
u8 throughout so that a width the header cannot express is
unrepresentable rather than silently truncated. The shard budget is no
longer capped at 255 either, since the codec carries up to 65536 and
validating a budget is left to it rather than repeated.

+++ crates/scuttlecast/src/proto/payload.rs
@@ -15,7 +15,7 @@ pub struct Hello {
     pub transfer_id: u64,
     pub block_size: NonZeroU32,
     pub blocks_per_slice: NonZeroU16,
-    pub parity_per_slice: u16,
+    pub parity_per_slice: u8,
     pub max_live_slices: NonZeroU16,
 }
